moved onSelect event handling to core plugin
