+++ lib/components/content.js
@@ -557,41 +557,7 @@ class Content extends React.Component {
     if (this.tmp.isCopying) return
     if (this.tmp.isComposing) return
 
-    let { state } = this.props
-    let { document, selection } = state
-    const native = window.getSelection()
-
-    if (!native.rangeCount) {
-      selection = selection.merge({ isFocused: false })
-      state = state.merge({ selection })
-      this.onChange(state)
-      return
-    }
-
-    // Calculate the Slate-specific selection based on the native one.
-    const { anchorNode, anchorOffset, focusNode, focusOffset } = native
-    const anchor = OffsetKey.findPoint(anchorNode, anchorOffset, state)
-    const focus = OffsetKey.findPoint(focusNode, focusOffset, state)
-
-    // If the native selection is inside text nodes, we can trust the native
-    // state and not need to re-render.
-    const isNative = (
-      anchorNode.nodeType == 3 &&
-      focusNode.nodeType == 3
-    )
-
-    state = state
-      .transform()
-      .focus()
-      .moveTo({
-        anchorKey: anchor.key,
-        anchorOffset: anchor.offset,
-        focusKey: focus.key,
-        focusOffset: focus.offset
-      })
-      .apply({ isNative })
-
-    this.onChange(state)
+    this.props.onSelect(e)
   }
 
   /**
@@ -731,4 +697,3 @@ class Content extends React.Component {
  */
 
 export default Content
-
