+++ lib/components/editor.js
@@ -28,6 +28,7 @@ class Editor extends React.Component {
     onDrop: React.PropTypes.func,
     onKeyDown: React.PropTypes.func,
     onPaste: React.PropTypes.func,
+    onSelect: React.PropTypes.func,
     onSelectionChange: React.PropTypes.func,
     placeholder: React.PropTypes.any,
     placeholderClassName: React.PropTypes.string,
@@ -202,6 +203,16 @@ class Editor extends React.Component {
   }
 
   /**
+   * On selection.
+   *
+   * @param {Mixed} ...args
+   */
+
+  onSelect = (...args) => {
+    this.onEvent('onSelect', ...args)
+  }
+
+  /**
    * Render the editor.
    *
    * @return {Element} element
@@ -217,6 +228,7 @@ class Editor extends React.Component {
         onDrop={this.onDrop}
         onKeyDown={this.onKeyDown}
         onPaste={this.onPaste}
+        onSelect={this.onSelect}
         readOnly={this.props.readOnly}
         renderMark={this.renderMark}
         renderNode={this.renderNode}
