+++ lib/plugins/core.js
@@ -5,6 +5,7 @@ import React from 'react'
 import String from '../utils/string'
 import keycode from 'keycode'
 import { IS_WINDOWS, IS_MAC } from '../utils/environment'
+import OffsetKey from '../utils/offset-key'
 
 /**
  * The default plugin.
@@ -306,6 +307,51 @@ function Plugin(options = {}) {
           return transform.apply()
         }
       }
+    },
+
+    /**
+     * The core `onSelect` handler.
+     *
+     * @param {Event} e
+     * @param {State} state
+     * @param {Editor} editor
+     * @return {State or Null}
+     */
+
+    onSelect(e, state, editor) {
+      let { document, selection } = state
+      const native = window.getSelection()
+
+      if (!native.rangeCount) {
+        selection = selection.merge({ isFocused: false })
+        state = state.merge({ selection })
+        return state
+      }
+
+      // Calculate the Slate-specific selection based on the native one.
+      const { anchorNode, anchorOffset, focusNode, focusOffset } = native
+      const anchor = OffsetKey.findPoint(anchorNode, anchorOffset, state)
+      const focus = OffsetKey.findPoint(focusNode, focusOffset, state)
+
+      // If the native selection is inside text nodes, we can trust the native
+      // state and not need to re-render.
+      const isNative = (
+        anchorNode.nodeType == 3 &&
+        focusNode.nodeType == 3
+      )
+
+      state = state
+        .transform()
+        .focus()
+        .moveTo({
+          anchorKey: anchor.key,
+          anchorOffset: anchor.offset,
+          focusKey: focus.key,
+          focusOffset: focus.offset
+        })
+        .apply({ isNative })
+
+      return state
     },
 
     /**
